python/biograph/requirements.txt
# Needed for installation:
setuptools>=49.2.0
# Biograph core:
pytabix
pysam
joblib
progressbar2
pyvcf
setproctitle
networkx
# Classifier:
numpy
pandas
# Scikit-learn models seem to be incompatible with previous versions, so pin to this version specifically.
scikit-learn==0.24.0
# Variant Databse (VDB):
pyarrow>=3.0.0,<4
pyathena>=2.1.1,<2.2.0
pyparsing>=2.4.7,<3
boto3>=1.14.0,<2
orjson>=3.3.0,<4
# for GRM (2.0.2 or later for determinism fix):
Truvari>=2.0.2
